@@ -1,152 +1,153 @@
 {
-    "name": "fundamental-styles",
-    "description": "Fundamental Library Styles is a Design System and HTML/CSS Component Library used to build modern Product User Experiences with SAP Fiori look and feel with any web technology. Learn more about this project at - http://sap.github.io/fundamental-styles/",
-    "engines": {
-        "yarn": ">= 4.10.3",
-        "node": ">= 18"
-    },
-    "packageManager": "yarn@4.10.3",
-    "scripts": {
-        "transform-stories": "ts-node stories-transformer.ts",
-        "build": "npx nx run-many --target=build --all",
-        "build:prod": "yarn run build --configuration=production --skip-nx-cache",
-        "build:theming-preview": "npx nx run theming-preview:build",
-        "build:visual-stories": "nx run-many --target=build-visual --all",
-        "lint:style": "stylelint \"packages/*/src/**/*.(sa|sc|c)ss\"",
-        "lint:js": "eslint . --ext .js",
-        "lint": "yarn run sync-entries && yarn run lint:style && yarn run lint:js",
-        "lint:fix": "yarn run sync-entries fix && yarn run lint:style --fix && yarn run lint:js --fix",
-        "lint:pre-commit": "echo running pre-commit lint... && yarn run lint && echo done!\n",
-        "hotfix-release": "node scripts/release-hotfix.js",
-        "size": "bundlesize --config config/bundlesize.json",
-        "start": "yarn run build:visual-stories && yarn run storybook:serve",
-        "start:prod": "yarn run storybook:serve:prod",
-        "test": "npx vitest --run",
-        "test:accessibility": "vitest accessibility --run",
-        "test:update": "npx nx run-many --target=update-storyshots --all",
-        "storybook": "yarn run storybook:serve",
-        "storybook:serve": "storybook dev -p 6006",
-        "storybook:serve:prod": "cross-env NODE_ENV=production yarn run storybook:prep && cross-env NODE_ENV=production yarn run start",
-        "storybook:ci": "yarn run storybook --ci --quiet",
-        "storybook:prep": "yarn run build:visual-stories",
-        "storybook:static": "yarn run storybook:prep && cross-env STORYBOOK_ENV=docs storybook build -c .storybook -o storybook-static --docs",
-        "storybook:chromatic": "yarn run storybook:prep && cross-env STORYBOOK_ENV=chromatic storybook build -c .storybook -o storybook-chromatic",
-        "prepare": "husky install",
-        "sync-fiori-components": "npx nx run styles:sync-entry",
-        "sync-entries": "npx nx run-many --target sync-entry --all",
-        "pre-commit": "lint-staged && yarn run lint:pre-commit"
-    },
-    "repository": {
-        "type": "git",
-        "url": "https://github.com/SAP/fundamental-styles.git"
-    },
-    "author": "",
-    "license": "Apache-2.0",
-    "homepage": "https://sap.github.io/fundamental-styles/",
-    "devDependencies": {
-        "@actions/core": "1.11.1",
-        "@babel/core": "7.23.9",
-        "@chromatic-com/storybook": "^4.1.1",
-        "@commitlint/cli": "18.6.0",
-        "@commitlint/config-conventional": "18.6.0",
-        "@jsdevtools/npm-publish": "3.0.1",
-        "@nx/devkit": "21.4.1",
-        "@nx/eslint": "21.4.1",
-        "@nx/eslint-plugin": "21.4.1",
-        "@nx/js": "21.4.1",
-        "@nx/plugin": "21.4.1",
-        "@nx/react": "21.4.1",
-        "@nx/web": "21.4.1",
-        "@sap-theming/theming-base-content": "11.32.1",
-        "@storybook/addon-a11y": "^9.1.3",
-        "@storybook/addon-docs": "^9.1.3",
-        "@storybook/addon-links": "^9.1.3",
-        "@storybook/html-vite": "^9.1.3",
-        "@storybook/icons": "^1.3.2",
-        "@swc/core": "1.5.7",
-        "@testing-library/react": "15.0.6",
-        "@types/fs-extra": "11.0.4",
-        "@types/node": "18.19.14",
-        "@types/react": "18.3.1",
-        "@types/react-dom": "18.3.0",
-        "@typescript-eslint/eslint-plugin": "7.18.0",
-        "@typescript-eslint/parser": "7.18.0",
-        "autoprefixer": "10.4.21",
-        "bundlesize": "0.18.2",
-        "chromatic": "^11.25.2",
-        "clean-css": "5.3.3",
-        "clean-css-cli": "5.4.1",
-        "colors": "1.4.0",
-        "conventional-changelog": "5.1.0",
-        "conventional-changelog-angular": "7.0.0",
-        "conventional-changelog-cli": "4.1.0",
-        "conventional-recommended-bump": "9.0.0",
-        "core-js": "3.35.1",
-        "cross-env": "7.0.3",
-        "cssnano": "7.0.5",
-        "eslint": "8.57.0",
-        "eslint-plugin-import": "2.31.0",
-        "eslint-plugin-jsx-a11y": "6.10.1",
-        "eslint-plugin-react": "7.31.11",
-        "eslint-plugin-react-hooks": "5.0.0",
-        "eslint-plugin-storybook": "^9.1.3",
-        "fast-glob": "3.3.2",
-        "fs-extra": "11.2.0",
-        "git-semver-tags": "4.1.1",
-        "glob": "8.0.3",
-        "husky": "7.0.4",
-        "jsonc-eslint-parser": "2.1.0",
-        "lerna": "8.2.1",
-        "lint-staged": "12.4.2",
-        "lodash": "4.17.21",
-        "markdown-to-jsx": "7.1.2",
-        "normalize.css": "8.0.1",
-        "npm-run-all": "4.1.5",
-        "nx": "21.4.1",
-        "nx-stylelint": "17.1.4",
-        "postcss": "8.4.38",
-        "postcss-banner": "4.0.1",
-        "postcss-cli": "9.1.0",
-        "postcss-import": "14.1.0",
-        "postcss-normalize-charset": "4.0.1",
-        "postcss-scss": "4.0.9",
-        "prettier": "^3.2.5",
-        "react": "18.3.1",
-        "react-dom": "18.3.1",
-        "react-is": "18.3.1",
-        "remark-gfm": "4.0.1",
-        "sass": "1.52.1",
-        "semver": "7.5.4",
-        "standard-version": "9.3.2",
-        "start-server-and-test": "1.14.0",
-        "storybook": "^9.1.3",
-        "stylelint": "16.8.2",
-        "stylelint-config-standard": "36.0.1",
-        "stylelint-config-standard-scss": "^13.1.0",
-        "stylelint-scss": "6.1.0",
-        "through2": "4.0.2",
-        "tocbot": "4.12.3",
-        "ts-node": "10.9.2",
-        "tslib": "2.3.0",
-        "typescript": "5.8.2",
-        "vite": "6.4.1",
-        "vite-tsconfig-paths": "4.2.0",
-        "vitest": "0.30.1",
-        "vitest-axe": "0.1.0"
-    },
-    "resolutions": {
-        "nx": "21.4.1"
-    },
-    "browserslist": [
-        "last 2 versions"
-    ],
-    "lint-staged": {
-        "{packages,stories,.storybook}/**/*.(sa|sc|c)ss": "stylelint --fix",
-        "*.js": "eslint --fix"
-    },
-    "dependencies": {
-        "@vitejs/plugin-react": "4.5.0",
-        "fast-deep-equal": "3.1.3",
-        "jsdom": "20.0.1"
-    }
+  "name": "fundamental-styles",
+  "description": "Fundamental Library Styles is a Design System and HTML/CSS Component Library used to build modern Product User Experiences with SAP Fiori look and feel with any web technology. Learn more about this project at - http://sap.github.io/fundamental-styles/",
+  "type": "module",
+  "engines": {
+    "yarn": ">= 4.10.3",
+    "node": ">= 22"
+  },
+  "packageManager": "yarn@4.10.3",
+  "scripts": {
+    "transform-stories": "ts-node stories-transformer.ts",
+    "build": "npx nx build workspace-plugins && npx nx run-many --target=build --all",
+    "build:prod": "npx nx build workspace-plugins --skip-nx-cache && npx nx run common-css:build --configuration=production --skip-nx-cache && npx nx run styles:build --configuration=production --skip-nx-cache && npx nx run cx:build --configuration=production --skip-nx-cache && npx nx run theming-preview:build --configuration=production --skip-nx-cache",
+    "build:theming-preview": "npx nx run theming-preview:build",
+    "build:visual-stories": "nx run-many --target=build-visual --all",
+    "lint:style": "stylelint \"packages/*/src/**/*.(sa|sc|c)ss\"",
+    "lint:js": "eslint . --ext .js",
+    "lint": "yarn run sync-entries && yarn run lint:style && yarn run lint:js",
+    "lint:fix": "yarn run sync-entries fix && yarn run lint:style --fix && yarn run lint:js --fix",
+    "lint:pre-commit": "echo running pre-commit lint... && yarn run lint && echo done!\n",
+    "hotfix-release": "node scripts/release-hotfix.js",
+    "size": "bundlesize --config config/bundlesize.json",
+    "start": "npx nx build workspace-plugins && yarn run build:visual-stories && yarn run storybook:serve",
+    "start:prod": "yarn run storybook:serve:prod",
+    "test": "npx vitest --run",
+    "test:accessibility": "vitest accessibility --run",
+    "test:update": "npx nx run-many --target=update-storyshots --all",
+    "storybook": "yarn run storybook:serve",
+    "storybook:serve": "storybook dev -p 6006",
+    "storybook:serve:prod": "cross-env NODE_ENV=production yarn run storybook:prep && cross-env NODE_ENV=production yarn run start",
+    "storybook:ci": "yarn run storybook --ci --quiet",
+    "storybook:prep": "npx nx build workspace-plugins && yarn run build:visual-stories",
+    "storybook:static": "yarn run storybook:prep && cross-env STORYBOOK_ENV=docs storybook build -c .storybook -o storybook-static --docs",
+    "storybook:chromatic": "yarn run storybook:prep && cross-env STORYBOOK_ENV=chromatic storybook build -c .storybook -o storybook-chromatic",
+    "prepare": "husky install",
+    "sync-fiori-components": "npx nx build workspace-plugins && npx nx run styles:sync-entry",
+    "sync-entries": "npx nx build workspace-plugins && npx nx run-many --target sync-entry --all",
+    "pre-commit": "lint-staged && yarn run lint:pre-commit"
+  },
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/SAP/fundamental-styles.git"
+  },
+  "author": "",
+  "license": "Apache-2.0",
+  "homepage": "https://sap.github.io/fundamental-styles/",
+  "devDependencies": {
+    "@actions/core": "1.11.1",
+    "@babel/core": "7.23.9",
+    "@chromatic-com/storybook": "^4.1.2",
+    "@commitlint/cli": "18.6.0",
+    "@commitlint/config-conventional": "18.6.0",
+    "@jsdevtools/npm-publish": "3.0.1",
+    "@nx/devkit": "21.4.1",
+    "@nx/eslint": "21.4.1",
+    "@nx/eslint-plugin": "21.4.1",
+    "@nx/js": "21.4.1",
+    "@nx/plugin": "21.4.1",
+    "@nx/react": "21.4.1",
+    "@nx/web": "21.4.1",
+    "@sap-theming/theming-base-content": "11.32.1",
+    "@storybook/addon-a11y": "^10.0.4",
+    "@storybook/addon-docs": "^10.0.4",
+    "@storybook/addon-links": "^10.0.4",
+    "@storybook/html-vite": "^10.0.4",
+    "@storybook/icons": "^1.3.2",
+    "@swc/core": "1.5.7",
+    "@testing-library/react": "15.0.6",
+    "@types/fs-extra": "11.0.4",
+    "@types/node": "18.19.14",
+    "@types/react": "18.3.1",
+    "@types/react-dom": "18.3.0",
+    "@typescript-eslint/eslint-plugin": "7.18.0",
+    "@typescript-eslint/parser": "7.18.0",
+    "autoprefixer": "10.4.21",
+    "bundlesize": "0.18.2",
+    "chromatic": "^11.25.2",
+    "clean-css": "5.3.3",
+    "clean-css-cli": "5.4.1",
+    "colors": "1.4.0",
+    "conventional-changelog": "5.1.0",
+    "conventional-changelog-angular": "7.0.0",
+    "conventional-changelog-cli": "4.1.0",
+    "conventional-recommended-bump": "9.0.0",
+    "core-js": "3.35.1",
+    "cross-env": "7.0.3",
+    "cssnano": "7.0.5",
+    "eslint": "8.57.0",
+    "eslint-plugin-import": "2.31.0",
+    "eslint-plugin-jsx-a11y": "6.10.1",
+    "eslint-plugin-react": "7.31.11",
+    "eslint-plugin-react-hooks": "5.0.0",
+    "eslint-plugin-storybook": "^10.0.4",
+    "fast-glob": "3.3.2",
+    "fs-extra": "11.2.0",
+    "git-semver-tags": "4.1.1",
+    "glob": "8.0.3",
+    "husky": "7.0.4",
+    "jsonc-eslint-parser": "2.1.0",
+    "lerna": "8.2.1",
+    "lint-staged": "12.4.2",
+    "lodash": "4.17.21",
+    "markdown-to-jsx": "7.1.2",
+    "normalize.css": "8.0.1",
+    "npm-run-all": "4.1.5",
+    "nx": "21.4.1",
+    "nx-stylelint": "17.1.4",
+    "postcss": "8.4.38",
+    "postcss-banner": "4.0.1",
+    "postcss-cli": "9.1.0",
+    "postcss-import": "14.1.0",
+    "postcss-normalize-charset": "4.0.1",
+    "postcss-scss": "4.0.9",
+    "prettier": "^3.2.5",
+    "react": "18.3.1",
+    "react-dom": "18.3.1",
+    "react-is": "18.3.1",
+    "remark-gfm": "4.0.1",
+    "sass": "1.52.1",
+    "semver": "7.5.4",
+    "standard-version": "9.3.2",
+    "start-server-and-test": "1.14.0",
+    "storybook": "^10.0.4",
+    "stylelint": "16.8.2",
+    "stylelint-config-standard": "36.0.1",
+    "stylelint-config-standard-scss": "^13.1.0",
+    "stylelint-scss": "6.1.0",
+    "through2": "4.0.2",
+    "tocbot": "4.12.3",
+    "ts-node": "10.9.2",
+    "tslib": "2.3.0",
+    "typescript": "5.8.2",
+    "vite": "6.4.1",
+    "vite-tsconfig-paths": "4.2.0",
+    "vitest": "0.30.1",
+    "vitest-axe": "0.1.0"
+  },
+  "resolutions": {
+    "nx": "21.4.1"
+  },
+  "browserslist": [
+    "last 2 versions"
+  ],
+  "lint-staged": {
+    "{packages,stories,.storybook}/**/*.(sa|sc|c)ss": "stylelint --fix",
+    "*.js": "eslint --fix"
+  },
+  "dependencies": {
+    "@vitejs/plugin-react": "4.5.0",
+    "fast-deep-equal": "3.1.3",
+    "jsdom": "20.0.1"
+  }
 }
