@@ -12,4 +12,4 @@ const Configuration = {
     }
 };
 
-module.exports = Configuration;
+export default Configuration;
