@@ -1,5 +1,5 @@
 import { addons } from 'storybook/manager-api';
-import fundamentalTheme from '../packages/storybook/src/lib/fundamentals';
+import fundamentalTheme from '../packages/storybook/src/lib/fundamentals.js';
 
 addons.setConfig({
   sidebar: {
